+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Added
 
+- prescribed slicing for `Import.names`
 - prescribed slicing for `Global.names`
 - prescribed slicing for `Nonlocal.names`